requirements.txt
Flask
Flask-Session
Flask-SQLAlchemy
requests
python-dotenv
datetime